+++ client_119.go
@@ -7,8 +7,6 @@
 
 package mail
 
-import "strings"
-
 // Send sends out the mail message
 func (c *Client) Send(messages ...*Msg) error {
 	if cerr := c.checkConn(); cerr != nil {
@@ -16,101 +14,9 @@ func (c *Client) Send(messages ...*Msg) error {
 	}
 	var errs []*SendError
 	for _, message := range messages {
-		message.sendError = nil
-		if message.encoding == NoEncoding {
-			if ok, _ := c.smtpClient.Extension("8BITMIME"); !ok {
-				sendErr := &SendError{Reason: ErrNoUnencoded, isTemp: false}
-				message.sendError = sendErr
-				errs = append(errs, sendErr)
-				continue
-			}
-		}
-		from, err := message.GetSender(false)
-		if err != nil {
-			sendErr := &SendError{Reason: ErrGetSender, errlist: []error{err}, isTemp: isTempError(err)}
-			message.sendError = sendErr
+		if sendErr := c.sendSingleMsg(message); sendErr != nil {
+			messages[id].sendError = sendErr
 			errs = append(errs, sendErr)
-			continue
-		}
-		rcpts, err := message.GetRecipients()
-		if err != nil {
-			sendErr := &SendError{Reason: ErrGetRcpts, errlist: []error{err}, isTemp: isTempError(err)}
-			message.sendError = sendErr
-			errs = append(errs, sendErr)
-			continue
-		}
-
-		if c.dsn {
-			if c.dsnmrtype != "" {
-				c.smtpClient.SetDSNMailReturnOption(string(c.dsnmrtype))
-			}
-		}
-		if err = c.smtpClient.Mail(from); err != nil {
-			sendErr := &SendError{Reason: ErrSMTPMailFrom, errlist: []error{err}, isTemp: isTempError(err)}
-			if resetSendErr := c.smtpClient.Reset(); resetSendErr != nil {
-				sendErr.errlist = append(sendErr.errlist, resetSendErr)
-			}
-			message.sendError = sendErr
-			errs = append(errs, sendErr)
-			continue
-		}
-		failed := false
-		rcptSendErr := &SendError{}
-		rcptSendErr.errlist = make([]error, 0)
-		rcptSendErr.rcpt = make([]string, 0)
-		rcptNotifyOpt := strings.Join(c.dsnrntype, ",")
-		c.smtpClient.SetDSNRcptNotifyOption(rcptNotifyOpt)
-		for _, rcpt := range rcpts {
-			if err = c.smtpClient.Rcpt(rcpt); err != nil {
-				rcptSendErr.Reason = ErrSMTPRcptTo
-				rcptSendErr.errlist = append(rcptSendErr.errlist, err)
-				rcptSendErr.rcpt = append(rcptSendErr.rcpt, rcpt)
-				rcptSendErr.isTemp = isTempError(err)
-				failed = true
-			}
-		}
-		if failed {
-			if resetSendErr := c.smtpClient.Reset(); resetSendErr != nil {
-				rcptSendErr.errlist = append(rcptSendErr.errlist, err)
-			}
-			message.sendError = rcptSendErr
-			errs = append(errs, rcptSendErr)
-			continue
-		}
-		writer, err := c.smtpClient.Data()
-		if err != nil {
-			sendErr := &SendError{Reason: ErrSMTPData, errlist: []error{err}, isTemp: isTempError(err)}
-			message.sendError = sendErr
-			errs = append(errs, sendErr)
-			continue
-		}
-		_, err = message.WriteTo(writer)
-		if err != nil {
-			sendErr := &SendError{Reason: ErrWriteContent, errlist: []error{err}, isTemp: isTempError(err)}
-			message.sendError = sendErr
-			errs = append(errs, sendErr)
-			continue
-		}
-		message.isDelivered = true
-
-		if err = writer.Close(); err != nil {
-			sendErr := &SendError{Reason: ErrSMTPDataClose, errlist: []error{err}, isTemp: isTempError(err)}
-			message.sendError = sendErr
-			errs = append(errs, sendErr)
-			continue
-		}
-
-		if err = c.Reset(); err != nil {
-			sendErr := &SendError{Reason: ErrSMTPReset, errlist: []error{err}, isTemp: isTempError(err)}
-			message.sendError = sendErr
-			errs = append(errs, sendErr)
-			continue
-		}
-		if err = c.checkConn(); err != nil {
-			sendErr := &SendError{Reason: ErrConnCheck, errlist: []error{err}, isTemp: isTempError(err)}
-			message.sendError = sendErr
-			errs = append(errs, sendErr)
-			continue
 		}
 	}
